requirements.txt
click==8.0.1
ecdsa==0.17.0
Flask==2.0.1
flask-oidc==1.4.0
httplib2==0.19.1
itsdangerous==2.0.1
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-jose==3.3.0
python-keycloak==0.25.0
rsa==4.7.2
Werkzeug==2.0.1
